+++ lib/cellular_info.dart
@@ -5,7 +5,18 @@ import 'package:cellular_info/model/nr_signal.dart';
 import 'cellular_info_platform_interface.dart';
 
 class CellularInfo {
+  static Future<List<SignalNr>> getCellInfo() async {
+    if (!Platform.isAndroid) {
+      throw UnsupportedError('cellular_info is only supported on Android.');
+    }
+    final list = await CellularInfoPlatform.instance.getCellInfo();
+    return list;
+  }
+
   static Future<List<SignalNr>> getNrCellInfo() async {
+    if (!Platform.isAndroid) {
+      throw UnsupportedError('cellular_info is only supported on Android.');
+    }
     final list = await CellularInfoPlatform.instance.getNrCellInfo();
     return list;
   }
@@ -19,6 +30,18 @@ class CellularInfo {
       return CellularInfoPlatform.instance.nrSignalStreamFromService;
     } else {
       return CellularInfoPlatform.instance.nrSignalStream;
+    }
+  }
+
+  /// [enableService] - if ture that android service start.
+  static Stream<List<SignalNr>> getAllCellInfoStream({bool enableService = false}) {
+    if (!Platform.isAndroid) {
+      throw UnsupportedError('cellular_info is only supported on Android.');
+    }
+    if (enableService) {
+      return CellularInfoPlatform.instance.nrSignalStreamFromService;
+    } else {
+      return CellularInfoPlatform.instance.allCellInfoStream;
     }
   }
 }