+++ lib/model/nr_signal.dart
@@ -1,7 +1,11 @@
 class SignalNr {
+  /// 0: 5g-NR,1: 4g-LTE.
+  int type = 0;
+  bool isRegistered = false;
   String? band;
   double? freq;
   int? arfcn;
+  int? pci;
   int? csiRsrp;
   int? csiRsrq;
   int? csiSinr;
@@ -11,9 +15,12 @@ class SignalNr {
   int? dbm;
 
   SignalNr({
+    this.type = 0,
+    this.isRegistered = false,
     this.band,
     this.freq,
     this.arfcn,
+    this.pci,
     this.csiRsrp,
     this.csiRsrq,
     this.csiSinr,
@@ -24,9 +31,12 @@ class SignalNr {
   });
 
   SignalNr.fromJson(Map<String, dynamic> json) {
+    type = json["type"];
+    isRegistered = json["isRegistered"];
     band = json["band"];
     freq = json["freq"];
     arfcn = json["arfcn"];
+    pci = json["pci"];
     csiRsrp = json['csiRsrp'];
     csiRsrq = json['csiRsrq'];
     csiSinr = json['csiSinr'];
@@ -38,9 +48,12 @@ class SignalNr {
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> map = {};
+    map["type"] = type;
+    map["isRegistered"] = isRegistered;
     map["band"] = band;
     map['freq'] = freq;
     map['arfcn'] = arfcn;
+    map["pci"] = pci;
     map['csiRsrp'] = csiRsrp;
     map['csiRsrq'] = csiRsrq;
     map['csiSinr'] = csiSinr;